Tentei fazer as transações da aplicação na memória.
