+++ Ateliex.DomainDrivenDesign.sln
@@ -40,7 +40,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ateliex.EntityFrameworkCore
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ateliex.Net.Http", "Ateliex.Net.Http\Ateliex.Net.Http.csproj", "{ECD8D626-23FB-44D1-9246-F6A71B5E0207}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ateliex.Web", "Ateliex.Web\Ateliex.Web.csproj", "{4C5CC470-685D-4153-803A-03364F544138}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ateliex.Web", "Ateliex.Web\Ateliex.Web.csproj", "{4C5CC470-685D-4153-803A-03364F544138}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ateliex.UI", "Ateliex.UI\Ateliex.UI.csproj", "{9F975D99-54B8-48DC-9102-E6CC635F2218}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ateliex.DTO", "Ateliex.DTO\Ateliex.DTO.csproj", "{9CC1FE5A-6D66-4573-A85C-6AA8CDCECC38}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -88,6 +92,14 @@ Global
 		{4C5CC470-685D-4153-803A-03364F544138}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4C5CC470-685D-4153-803A-03364F544138}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4C5CC470-685D-4153-803A-03364F544138}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9F975D99-54B8-48DC-9102-E6CC635F2218}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9F975D99-54B8-48DC-9102-E6CC635F2218}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9F975D99-54B8-48DC-9102-E6CC635F2218}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9F975D99-54B8-48DC-9102-E6CC635F2218}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9CC1FE5A-6D66-4573-A85C-6AA8CDCECC38}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9CC1FE5A-6D66-4573-A85C-6AA8CDCECC38}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9CC1FE5A-6D66-4573-A85C-6AA8CDCECC38}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9CC1FE5A-6D66-4573-A85C-6AA8CDCECC38}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -106,6 +118,8 @@ Global
 		{3C3FB54C-E844-45AB-960D-8373A14646FD} = {89E4BB85-AE46-4E96-93C7-269794163702}
 		{ECD8D626-23FB-44D1-9246-F6A71B5E0207} = {89E4BB85-AE46-4E96-93C7-269794163702}
 		{4C5CC470-685D-4153-803A-03364F544138} = {BAEF6DF0-4906-4E64-9158-B00BDC3A0BD2}
+		{9F975D99-54B8-48DC-9102-E6CC635F2218} = {BAEF6DF0-4906-4E64-9158-B00BDC3A0BD2}
+		{9CC1FE5A-6D66-4573-A85C-6AA8CDCECC38} = {89E4BB85-AE46-4E96-93C7-269794163702}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {C848A1A2-5716-4D07-8AA9-C41C4004F825}
