+++ Ateliex.Infrastructure.Client/Cadastro/Modelos/ModelosInfraService.cs
@@ -7,13 +7,17 @@ using System.Threading.Tasks;
 
 namespace Ateliex.Cadastro.Modelos
 {
-    public class ModelosService : IConsultaDeModelos, IRepositorioDeModelos
+    public class ModelosInfraService : IConsultaDeModelos, IRepositorioDeModelos
     {
         private readonly ModelosDbService db;
 
-        public ModelosService(ModelosDbService db)
+        private readonly ModelosHttpService http;
+
+        public ModelosInfraService(ModelosDbService db, ModelosHttpService http)
         {
             this.db = db;
+
+            this.http = http;
         }
 
         public async Task<RespostaDeConsultaDeModelos> ConsultaModelos(ParametrosDeConsultaDeModelos parametros)
@@ -26,16 +30,29 @@ namespace Ateliex.Cadastro.Modelos
         public async Task Add(Modelo modelo)
         {
             await db.Add(modelo);
+
+            var solicitacao = new SolicitacaoDeCadastroDeModelo
+            {
+                Nome = modelo.Nome
+            };
+
+            await http.CadastraModelo(solicitacao);
         }
 
         public async Task Update(Modelo modelo)
         {
             await db.Update(modelo);
+
+            //await http.AdicionaRecursoDeModelo(solicitacao);
+
+            // TODO.
         }
 
         public async Task Remove(Modelo modelo)
         {
             await db.Remove(modelo);
+
+            await http.RemoveModelo(modelo.Codigo);
         }
 
         public async Task<Modelo> ObtemModelo(string id)
@@ -45,9 +62,9 @@ namespace Ateliex.Cadastro.Modelos
             return modelo;
         }
 
-        public async Task<IEnumerable<Modelo>> ObtemObservavelDeModelos()
+        public async Task<Modelo[]> ObtemModelos()
         {
-            var modelos = await db.ObtemObservavelDeModelos();
+            var modelos = await db.ObtemModelos();
 
             return modelos;
         }
