+++ Ateliex.Infrastructure.Client/Cadastro/Modelos/ModelosPackage.cs
@@ -7,7 +7,7 @@ namespace Ateliex.Cadastro.Modelos
     {
         public void RegisterServices(IServiceCollection services)
         {
-            services.AddTransient<ModelosService>();
+            services.AddTransient<ModelosInfraService>();
 
             services.AddTransient<ModelosDbService>();
