+++ Ateliex.Infrastructure.Client/Decisoes/Comerciais/PlanosComerciaisPackage.cs
@@ -6,7 +6,7 @@ namespace Ateliex.Decisoes.Comerciais
     {
         public void RegisterServices(IServiceCollection services)
         {
-            services.AddTransient<PlanosComerciaisService>();
+            services.AddTransient<PlanosComerciaisInfraService>();
 
             services.AddTransient<PlanosComerciaisDbService>();
