+++ Ateliex.Web/Startup.cs
@@ -2,6 +2,9 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using System.Transactions;
+using Ateliex.Cadastro.Modelos;
+using Ateliex.Decisoes.Comerciais;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.HttpsPolicy;
@@ -30,6 +33,20 @@ namespace Ateliex
                 options.UseSqlServer(Configuration.GetConnectionString("DefaultConnection")));
 
             services.AddControllers();
+
+            services.AddTransient(typeof(IUnitOfWork), typeof(TransactionScopeManager));
+
+            services.AddTransient<ICadastroDeModelos, CadastroDeModelos>();
+
+            services.AddTransient<IConsultaDeModelos, ModelosDbService>();
+
+            services.AddTransient<IRepositorioDeModelos, ModelosDbService>();
+
+            services.AddTransient<IConsultaDePlanosComerciais, PlanosComerciaisDbService>();
+
+            services.AddTransient<IRepositorioDePlanosComerciais, PlanosComerciaisDbService>();
+
+            services.AddTransient<IPlanejamentoComercial, PlanejamentoComercial>();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
